requirements.txt
alembic==1.7.7
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopage==0.5.0
backcall==0.2.0
bleach==4.1.0
bottle==0.12.19
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorlog==6.6.0
cycler==0.11.0
dash==2.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
deap==1.3.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
Flask-Compress==1.11
fonttools==4.29.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mido==1.2.10
mistune==0.8.4
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.3
optuna==2.10.0
optuna-dashboard==0.6.1
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.32
stack-data==0.2.0
stevedore==3.5.0
stopit==1.1.2
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.11.0
tornado==6.1
TPOT==0.11.7
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
xgboost==1.5.2
zipp==3.7.0
